pyproject.toml
[project]
name = "blog-post-generato-backend"
version = "0.1.0"
description = "AI Blog Generator Backend with FastAPI, MongoDB, and JWT Authentication"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain-core>=0.3.75",
    "langgraph>=0.6.6",
    "motor>=3.7.1",
    "openai>=1.105.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.7",
    "pymongo>=4.14.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "tavily-python>=0.7.11",
    "uvicorn>=0.35.0",
]
